requirements.txt
beautifulsoup4==4.10.0
chart_studio==1.1.0
cufflinks==0.17.3
pystan==2.19.1.1
fbprophet==0.7.1
keras==2.8.0
keras_nightly==2.5.0.dev2021032900
lxml==4.7.1
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
plotly==4.1.1
requests==2.27.1
seaborn==0.11.1
spacy==3.2.4
streamlit==0.49.0
yfinance==0.1.54
